Fix warnings caused in Unit Tests
These were originally caught by @ypatia in upgrading Catch2 library for
2.3. These are needed for backporting to work with GCC 10.2.

+++ test/src/unit-filter-pipeline.cc
@@ -1716,7 +1716,7 @@ TEST_CASE("Filter: Test bit width reduction", "[filter]") {
       uint64_t elt = 0;
       CHECK(chunked_buffer.read(&elt, sizeof(uint64_t), (i * sizeof(uint64_t)))
                 .ok());
-      CHECK(elt == rng(gen_copy));
+      CHECK((int64_t)elt == rng(gen_copy));
     }
   }
